app/services/layout_analysis/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.21.0
opencv-python>=4.8.0
pillow>=9.5.0
numpy>=1.21.0
# paddlepaddle-gpu>=2.5.0
paddleocr>=2.7.0
easyocr>=1.7.0
pytesseract>=0.3.10
datasets>=2.14.0
accelerate>=0.21.0
sentencepiece>=0.1.99
scikit-learn>=1.3.0
scipy>=1.10.0
pdf2image>=1.16.0